+++ src/main/java/org/embl/mobie/io/n5/writers/WriteImagePlusToN5.java
@@ -41,17 +41,17 @@ import java.util.Map;
 // splits functionality from https://github.com/bigdataviewer/bigdataviewer_fiji/blob/master/src/main/java/bdv/ij/ExportImagePlusAsN5PlugIn.java
 // separating from GUI. Provides general export functions for ImagePlus.
 
-public class WriteImgPlusToN5 {
+public class WriteImagePlusToN5 {
 
     // export, generating default source transform, and default resolutions / subdivisions
     public void export(ImagePlus imp, String xmlPath, DownsampleBlock.DownsamplingMethod downsamplingMethod,
                        Compression compression) {
-        if (!WriteImgPlusToN5Helper.isImageSuitable(imp)) {
+        if (!WriteImagePlusToN5Helper.isImageSuitable(imp)) {
             return;
         }
 
-        FinalVoxelDimensions voxelSize = WriteImgPlusToN5Helper.getVoxelSize(imp);
-        final AffineTransform3D sourceTransform = WriteImgPlusToN5Helper.generateSourceTransform(voxelSize);
+        FinalVoxelDimensions voxelSize = WriteImagePlusToN5Helper.getVoxelSize(imp);
+        final AffineTransform3D sourceTransform = WriteImagePlusToN5Helper.generateSourceTransform(voxelSize);
 
         Parameters defaultParameters = generateDefaultParameters(imp, xmlPath, sourceTransform, downsamplingMethod,
                 compression, null);
@@ -62,7 +62,7 @@ public class WriteImgPlusToN5 {
     // export, generating default resolutions / subdivisions
     public void export(ImagePlus imp, String xmlPath, AffineTransform3D sourceTransform,
                        DownsampleBlock.DownsamplingMethod downsamplingMethod, Compression compression) {
-        if (!WriteImgPlusToN5Helper.isImageSuitable(imp)) {
+        if (!WriteImagePlusToN5Helper.isImageSuitable(imp)) {
             return;
         }
 
@@ -76,7 +76,7 @@ public class WriteImgPlusToN5 {
     public void export(ImagePlus imp, String xmlPath, AffineTransform3D sourceTransform,
                        DownsampleBlock.DownsamplingMethod downsamplingMethod, Compression compression,
                        String[] viewSetupNames) {
-        if (!WriteImgPlusToN5Helper.isImageSuitable(imp)) {
+        if (!WriteImagePlusToN5Helper.isImageSuitable(imp)) {
             return;
         }
 
@@ -113,12 +113,12 @@ public class WriteImgPlusToN5 {
         String seqFilename = xmlPath;
         if (!seqFilename.endsWith(".xml"))
             seqFilename += ".xml";
-        final File seqFile = WriteImgPlusToN5Helper.getSeqFileFromPath(seqFilename);
+        final File seqFile = WriteImagePlusToN5Helper.getSeqFileFromPath(seqFilename);
         if (seqFile == null) {
             return;
         }
 
-        final File n5File = WriteImgPlusToN5Helper.getN5FileFromXmlPath(seqFilename);
+        final File n5File = WriteImagePlusToN5Helper.getN5FileFromXmlPath(seqFilename);
 
         Parameters exportParameters = new Parameters(resolutions, subdivisions, seqFile, n5File, sourceTransform,
                 downsamplingMethod, compression, viewSetupNames);
@@ -129,8 +129,8 @@ public class WriteImgPlusToN5 {
     protected Parameters generateDefaultParameters(ImagePlus imp, String xmlPath, AffineTransform3D sourceTransform,
                                                    DownsampleBlock.DownsamplingMethod downsamplingMethod, Compression compression,
                                                    String[] viewSetupNames) {
-        FinalVoxelDimensions voxelSize = WriteImgPlusToN5Helper.getVoxelSize(imp);
-        FinalDimensions size = WriteImgPlusToN5Helper.getSize(imp);
+        FinalVoxelDimensions voxelSize = WriteImagePlusToN5Helper.getVoxelSize(imp);
+        FinalDimensions size = WriteImagePlusToN5Helper.getSize(imp);
 
         // propose reasonable mipmap settings
         final int maxNumElements = 64 * 64 * 64;
@@ -149,12 +149,12 @@ public class WriteImgPlusToN5 {
         String seqFilename = xmlPath;
         if (!seqFilename.endsWith(".xml"))
             seqFilename += ".xml";
-        final File seqFile = WriteImgPlusToN5Helper.getSeqFileFromPath(seqFilename);
+        final File seqFile = WriteImagePlusToN5Helper.getSeqFileFromPath(seqFilename);
         if (seqFile == null) {
             return null;
         }
 
-        final File n5File = WriteImgPlusToN5Helper.getN5FileFromXmlPath(seqFilename);
+        final File n5File = WriteImagePlusToN5Helper.getN5FileFromXmlPath(seqFilename);
 
         return new Parameters(resolutions, subdivisions, seqFile, n5File, sourceTransform,
                 downsamplingMethod, compression, viewSetupNames);
@@ -162,8 +162,8 @@ public class WriteImgPlusToN5 {
 
     protected void export(ImagePlus imp, Parameters params) {
 
-        FinalVoxelDimensions voxelSize = WriteImgPlusToN5Helper.getVoxelSize(imp);
-        FinalDimensions size = WriteImgPlusToN5Helper.getSize(imp);
+        FinalVoxelDimensions voxelSize = WriteImagePlusToN5Helper.getVoxelSize(imp);
+        FinalDimensions size = WriteImagePlusToN5Helper.getSize(imp);
 
         final ProgressWriter progressWriter = new ProgressWriterIJ();
         progressWriter.out().println("starting export...");
